pyproject.toml
[project]
name = "explicitlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "boto3>=1.40.45",
    "deepspeed>=0.18.0",
    "dvc[s3]>=3.63.0",
    "hydra-zen>=0.16.0",
    "hydra-core>=1.3.0",
    "mpi4py>=4.1.1",
    "setproctitle>=1.3.0",
    "swanlab>=0.6.11",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "dotenv>=0.9.9",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
